# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comet API",
    "version": "1.0.0",
    "description": "API documentation for Comet: The Speed of Light Comic Reader."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API"
    }
  ],
  "components": {
    "schemas": {
...
      "FriendDTO": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "image"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "FriendRequestDTO": {
        "type": "object",
        "required": [
          "id",
          "senderId",
          "receiverId",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "receiverId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "DECLINED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "sender": {
            "$ref": "#/components/schemas/FriendDTO"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/friends": {
      "get": {
        "summary": "List friends",
        "tags": [
          "Friends"
        ],
        "responses": {
          "200": {
            "description": "List of friends",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "get_friends"
      },
      "delete": {
        "summary": "Remove a friend",
        "tags": [
          "Friends"
        ],
        "parameters": [
          {
            "name": "friendId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Friend removed"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "delete_friends"
      }
    },
    "/friends/invite": {
      "post": {
        "summary": "Invite a friend by email",
        "tags": [
          "Friends"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite sent"
          },
          "404": {
            "description": "User not found"
          }
        },
        "operationId": "post_friends_invite"
      }
    },
    "/friends/requests": {
      "get": {
        "summary": "List friend requests",
        "tags": [
          "Friends"
        ],
        "responses": {
          "200": {
            "description": "List of requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendRequestDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "get_friends_requests"
      },
      "post": {
        "summary": "Send a friend request",
        "tags": [
          "Friends"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "receiverId"
                ],
                "properties": {
                  "receiverId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request sent"
          },
          "400": {
            "description": "Request already exists"
          }
        },
        "operationId": "post_friends_requests"
      }
    },
    "/friends/requests/{id}": {
      "put": {
        "summary": "Accept or decline friend request",
        "tags": [
          "Friends"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "ACCEPT",
                      "DECLINE"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request processed"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "put_friends_requests_id"
      }
    },
...
  }
}